+++ modules/core/src/main/java/org/apache/ignite/internal/processors/platform/client/ClientRequest.java
@@ -25,7 +25,7 @@ import org.apache.ignite.internal.processors.odbc.SqlListenerRequest;
  */
 public class ClientRequest extends SqlListenerRequest {
     /** Request id. */
-    private final int requestId;
+    private final long requestId;
 
     /**
      * Ctor.
@@ -33,7 +33,7 @@ public class ClientRequest extends SqlListenerRequest {
      * @param reader Reader.
      */
     public ClientRequest(BinaryRawReader reader) {
-        requestId = reader.readInt();
+        requestId = reader.readLong();
     }
 
     /**
@@ -41,7 +41,7 @@ public class ClientRequest extends SqlListenerRequest {
      *
      * @return Data.
      */
-    public int getRequestId() {
+    public long getRequestId() {
         return requestId;
     }
